Tests/LibWeb/Text/expected/display_list/contenteditable-caret-move-invalidation.txt

Before:   FillRect@1 rect=[8,8 1x18] color=rgb(0, 0, 0)
After:   FillRect@1 rect=[71,8 1x18] color=rgb(0, 0, 0)
